moved user model
