+++ src/modules/file/file.repository.ts
@@ -16,7 +16,7 @@ import {
   PrimaryKey,
   Table,
 } from 'sequelize-typescript';
-import { UserModel } from '../user/user.repository';
+import { UserModel } from '../user/user.model';
 import { User } from '../user/user.domain';
 import { Folder, FolderAttributes } from '../folder/folder.domain';
 import { Pagination } from '../../lib/pagination';
