+++ src/modules/keyserver/key-server.repository.ts
@@ -10,7 +10,7 @@ import {
   Table,
 } from 'sequelize-typescript';
 import { UserAttributes } from '../user/user.domain';
-import { UserModel } from '../user/user.repository';
+import { UserModel } from '../user/user.model';
 import { KeyServer, KeyServerAttributes } from './key-server.domain';
 
 @Table({
