+++ src/modules/share/share.module.ts
@@ -7,8 +7,9 @@ import { FileModel, SequelizeFileRepository } from '../file/file.repository';
 import { FolderModel } from '../folder/folder.model';
 import { FolderModule } from '../folder/folder.module';
 import { SequelizeFolderRepository } from '../folder/folder.repository';
+import { UserModel } from '../user/user.model';
 import { UserModule } from '../user/user.module';
-import { SequelizeUserRepository, UserModel } from '../user/user.repository';
+import { SequelizeUserRepository } from '../user/user.repository';
 import { ShareController } from './share.controller';
 import { SequelizeShareRepository, ShareModel } from './share.repository';
 import { ShareUseCases } from './share.usecase';
